+++ src/main/java/com/umc/hwaroak/controller/MemberController.java
@@ -1,8 +1,10 @@
 package com.umc.hwaroak.controller;
 
 
+import com.umc.hwaroak.dto.response.EmotionSummaryResponseDto;
 import com.umc.hwaroak.dto.response.MemberResponseDto;
 import com.umc.hwaroak.dto.request.MemberRequestDto;
+import com.umc.hwaroak.service.EmotionSummaryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -12,6 +14,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import com.umc.hwaroak.service.MemberService;
 
+import java.time.YearMonth;
 import java.util.List;
 
 @Tag(name = "Member API", description = "사용자 관련 API")
@@ -61,6 +64,30 @@ public class MemberController {
             @PathVariable Long itemId
     ){
         return memberService.changeSelectedItem(itemId);
+    }
+
+    private final EmotionSummaryService emotionSummaryService;
+
+    @GetMapping("emotions/preview")
+    @Operation(summary = "이번달 감정분석 preview 조회",
+            description = "이번달 감정 카테고리별 개수와 비율을 조회합니다. 반올림 때문에 비율 총합이 100이 넘을 수도 있습니다. 분석할 데이터가 없는 경우 null이 반환됩니다.")
+    @ApiResponse(content = @Content(schema = @Schema(implementation = MemberResponseDto.PreviewDto.class)))
+    public MemberResponseDto.PreviewDto getEmotionSummary() {
+
+        // 오늘 날짜 기준으로 이번 달 구하기
+        String yearMonth = YearMonth.now().toString(); // ex. "2025-07"
+
+        return emotionSummaryService.getPreviewEmotionSummary(yearMonth);
+    }
+
+    @GetMapping("emotions/{summaryMonth}")
+    @Operation(summary = "감정분석 상세 조회", description = "특정달 감정분석을 조회합니다.")
+    @ApiResponse(content = @Content(schema = @Schema(implementation = MemberResponseDto.DetailDto.class)))
+    public MemberResponseDto.DetailDto getDetailEmotionSummary(
+            @Schema(description = "조회할 연월", example = "2025-07")
+            @PathVariable String summaryMonth
+    ) {
+        return emotionSummaryService.getDetailEmotionSummary(summaryMonth);
     }
 
 }
