+++ src/main/java/com/umc/hwaroak/serviceImpl/EmotionSummaryServiceImpl.java
@@ -7,6 +7,7 @@ import com.umc.hwaroak.domain.Member;
 import com.umc.hwaroak.domain.common.Emotion;
 import com.umc.hwaroak.domain.common.EmotionCategory;
 import com.umc.hwaroak.dto.response.EmotionSummaryResponseDto;
+import com.umc.hwaroak.dto.response.MemberResponseDto;
 import com.umc.hwaroak.repository.DiaryRepository;
 import com.umc.hwaroak.repository.EmotionSummaryRepository;
 import com.umc.hwaroak.service.EmotionSummaryService;
@@ -36,7 +37,7 @@ public class EmotionSummaryServiceImpl implements EmotionSummaryService {
 
     @Override
     @Transactional(readOnly=true)
-    public EmotionSummaryResponseDto.PreviewDto getPreviewEmotionSummary(String yearMonth) {
+    public MemberResponseDto.PreviewDto getPreviewEmotionSummary(String yearMonth) {
 
         // 멤버ID 받아와서 감정분석 데이터 조회
         Long memberId = memberLoader.getCurrentMemberId();
@@ -46,14 +47,14 @@ public class EmotionSummaryServiceImpl implements EmotionSummaryService {
 
         if (summary == null) {
             log.info("감정 요약이 존재하지 않습니다. memberId = {}, yearMonth = {}", memberId, yearMonth);
-            return new EmotionSummaryResponseDto.PreviewDto(); // 비어 있는 감정 통계 응답
+            return new MemberResponseDto.PreviewDto(); // 비어 있는 감정 통계 응답
         }
 
         return createPreviewFromSummary(summary);
     }
 
     // 감정 카테고리별 통계 정보 PreviewDto에 담아 반환
-    private EmotionSummaryResponseDto.PreviewDto createPreviewFromSummary(EmotionSummary summary) {
+    private MemberResponseDto.PreviewDto createPreviewFromSummary(EmotionSummary summary) {
 
         // 감정 전체 개수 카운트
         int total = summary.getCalmCount() + summary.getHappyCount()
@@ -64,13 +65,13 @@ public class EmotionSummaryServiceImpl implements EmotionSummaryService {
         }
 
         // 개수 기반으로 비율 계산하고 두 필드 모두 리턴
-        Map<EmotionCategory, EmotionSummaryResponseDto.EmotionCount> map = new EnumMap<>(EmotionCategory.class);
-        map.put(EmotionCategory.CALM, new EmotionSummaryResponseDto.EmotionCount(summary.getCalmCount(), calculatePercent(summary.getCalmCount(), total)));
-        map.put(EmotionCategory.HAPPY, new EmotionSummaryResponseDto.EmotionCount(summary.getHappyCount(), calculatePercent(summary.getHappyCount(), total)));
-        map.put(EmotionCategory.SAD, new EmotionSummaryResponseDto.EmotionCount(summary.getSadCount(), calculatePercent(summary.getSadCount(), total)));
-        map.put(EmotionCategory.ANGRY, new EmotionSummaryResponseDto.EmotionCount(summary.getAngryCount(), calculatePercent(summary.getAngryCount(), total)));
+        Map<EmotionCategory, MemberResponseDto.EmotionCount> map = new EnumMap<>(EmotionCategory.class);
+        map.put(EmotionCategory.CALM, new MemberResponseDto.EmotionCount(summary.getCalmCount(), calculatePercent(summary.getCalmCount(), total)));
+        map.put(EmotionCategory.HAPPY, new MemberResponseDto.EmotionCount(summary.getHappyCount(), calculatePercent(summary.getHappyCount(), total)));
+        map.put(EmotionCategory.SAD, new MemberResponseDto.EmotionCount(summary.getSadCount(), calculatePercent(summary.getSadCount(), total)));
+        map.put(EmotionCategory.ANGRY, new MemberResponseDto.EmotionCount(summary.getAngryCount(), calculatePercent(summary.getAngryCount(), total)));
 
-        return new EmotionSummaryResponseDto.PreviewDto(map);
+        return new MemberResponseDto.PreviewDto(map);
     }
 
     // 비율 계산 메소드 - 소수점 첫째 자리까지 반올림
@@ -79,7 +80,7 @@ public class EmotionSummaryServiceImpl implements EmotionSummaryService {
     }
 
     @Override
-    public EmotionSummaryResponseDto.DetailDto getDetailEmotionSummary(String yearMonth) {
+    public MemberResponseDto.DetailDto getDetailEmotionSummary(String yearMonth) {
 
         // 멤버ID 받아와서 감정분석 데이터 조회
         Long memberId = memberLoader.getCurrentMemberId();
@@ -89,10 +90,10 @@ public class EmotionSummaryServiceImpl implements EmotionSummaryService {
 
         if (summary == null) {
             log.info("감정 요약이 존재하지 않습니다. memberId = {}, yearMonth = {}", memberId, yearMonth);
-            return new EmotionSummaryResponseDto.DetailDto(); // 비어 있는 감정 통계 응답
+            return new MemberResponseDto.DetailDto(); // 비어 있는 감정 통계 응답
         }
 
-        return EmotionSummaryResponseDto.DetailDto.builder()
+        return MemberResponseDto.DetailDto.builder()
                 .diaryCount(summary.getDiaryCount())
                 .emotionSummary(createPreviewFromSummary(summary))
                 .message(summary.getSummaryMessage())
